requirements/common.txt
amqp==2.5.0
arrow==0.14.2
asn1crypto==0.24.0
Babel==2.7.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bibtexparser==1.1.0
billiard==3.5.0.5
binaryornot==0.4.4
blinker==1.4
bsddb3==6.2.6
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
cryptography==2.7
Cython==0.29.10
fabric==2.4.0
filedepot==0.6.0
Flask==1.0.3
#Flask-Admin==1.5.3
#Flask-BabelEx==0.9.3
Flask-Login==0.3.2
#Flask-Mail==0.9.1
#Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-WTF==0.14.2
future==0.17.1
html5lib==1.0.1
idna==2.8
ijson==2.4
invoke==1.2.0
isodate==0.6.0
itsdangerous==1.1.0
javabridge==1.0.18
Jinja2==2.10.1
jinja2-time==0.2.0
keepalive==0.5
kombu==4.4.0
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.3
nose==1.3.7
numpy==1.16.4
pandas==0.24.2
paramiko==2.5.0
passlib==1.7.1
poyo==0.4.2
pycparser==2.19
pygeoif==0.7
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyshp==2.1.0
python-dateutil==2.8.0
python-magic==0.4.15
python-slugify==3.0.2
pytidylib==0.3.2
pytz==2019.1
#git+git://github.com/RDFLib/rdflib.git#egg=rdflib
git+git://github.com/gordom6/rdflib.git#egg=rdflib
git+git://github.com/RDFLib/rdflib-jsonld.git#egg=rdflib-jsonld
git+git://github.com/tetherless-world/markdown-rdfa.git#egg=markdown-rdfa
redis==3.2.0
requests[security]==2.22.0
requests-testadapter==0.3.0
sadi==0.5.7
scipy==1.2.1
setlr==0.2.11
six==1.12.0
soupsieve==1.9.1
SPARQLWrapper==1.8.4
speaklater==1.3
text-unidecode==1.2
toposort==1.5
Unidecode==1.0.23
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==0.14.1
whichcraft==0.5.2
WTForms==2.2.1
xlrd==1.2.0
elasticsearch
